Start Move To Flutter
After careful consideration, I've decided to move frontend development
to flutter instead of natively developing both iOS and Android apps.
Mostly, this is due to time and work concerns.  Furthermore, Flutter has
an enforced way of managing state, which Swift UI does not benefit from.
Flutter also benefits from crossplatform development and tight Firebase
integration.

This commit only contains Flutter code for a login flow to a homepage
with a button that brings you to a blank channel page.  Further commits
will flesh out this Flutter app to the level of development the SwiftUI
app was at before the SwiftUI meteor-ios directory will be deleted.
This commit also has SwiftUI code for meteor-ios that was in progress
before I switched to Flutter.