Write documentation for new backend
Started writing documentation for services in the new backend.  The
new backend is planned on being a containerized Node application with
module-level microservices.  This will be
replacing Google Cloud Functions so I can access everything I need
through API endpoints with a known structure.  It also opens the
possibility for a move to container-level microservices in the
near or far future.