Add more debugging statements
I'm getting the same error.  I'm also looking into setting up PubSub as a message bus for Mux events to make this a little easier to deal with.  Setting up PubSub with this would also make sense if I plan to create my own transcoding pipeline, as that's a highly asynchronous task.