Update video query definition to use numbers
The video query function was returning no results for "before" and "after" fields because I was passing it strings.  While the TS definition for IVideoQuery specifies numbers (so the TS type checker didn't catch me passing strings), the req.query type of 'any' meant that the value was simply casted as a number without regard to it being a string.  I fixed this by changing the IVideoQuery definition to use strings, and explicitly parsing an integer out of the passed values.