Add ready videos to search index
VideoContentService was failing to add videos to the search index upon completing their transcoding job.

I added the required call to SearchService.  I also added a system of elevated permissions for server-initiated service calls through the `elevated` property on `IServiceInvocationContext`.  Not all services check for `elevated`, so *use with caution*.