Add finished v0.2.0 web app
This is the roughly finished v0.2.0 web app.  It is subject to future change before 0.2, but as of now it's in its finished form. 

There are many things missing from this page however.  At some point after phase 2, the following things need implementation: SEO (auth bypass and all), analytics integration, accessibility features, more video info (upload date), zero states, and error states.  This is not an exhaustive list.  There are many things wrong with the site in this commit, but all the important features (CRUD operations + video viewing) are implemented.