Change watch time to store intentionally duplicated data
It's better to store both a purposefully duplicated counter (each time a user watches a video, the time is counted without deduplication) for calculating a user's payment distribution and a deduplicated counter for ranking videos.  The duplicated counter exists for each user watching a video, and is reset every billing cycle.  The deduplicated counter exists only to count total watch time on a video, and persists through billing cycles.