Add deletion functionality to W.T.
Added delete methods for both deleting a single user out of the watch time table and an entire video.  The video deletion could take multiple seconds or even minutes, so I added a "deleted" flag to prevent modifications during the deletion process (required by the SCAN command).  Single user deletion will not be used in phase 3, but instead be used to reset a user's watch time at the end of every billing cycle in phase 4.