Add watch time data to mobile UI
Video titles are shown underneath the player (similar to YT), underneath which the watch time information sits.  Minor changes to the util lib were made to make it easier to format the watch time, and an additional service call for watch time getting was added.  The models were updated with the watch time and duration properties.