Transfer infrastructure to Terraform's control
All infrastructure has been transferred to Terraform except authentication.  Auth0 is currently configured by Terraform, but the backend and frontend are still using the old tenant.  Future steps for this are 1) update backend to read JWT parameters from environment variables (and update the Helm chart as needed) and 2) update frontend apps (mobile + web) to use the new auth0 domain.

Something that also needs to be done is an update to the github CI.  The only thing it needs to do now is build docker images, since application deployment is also handled by Terraform with the Helm provider.

An important note with this commit is that the previously vague distinction between production and development naming has been erased.  Anything visible under https://swish.tv is now labelled as production.