Finish last of Terraform setup
Terraform has streamlined the deployment workflow, but the only thing missing is a link between the Github build results (container images) and the Helm release.  Since the most recent image is tagged with the most recent branch SHA, I attempted to use the Github provider to fetch the most recent HEAD SHA.  However, due to (presumably) a limitation in the Github API, I must put the Swish repo into an organization to use this functionality.  If I were to put it in an org, the benefits I get from student pro would leave the repo, so for now I'm forgoing this linking step and manually updating the terraform variable file every build.

Closes #42