Fix the credentials exchange hook
Before this change, the hook removed any scopes granted by Auth0 to the token during the exchange.  This wasn't a problem with the M to M test agent, as Swish uses no scopes.  However since the hook also runs on the Terraform agent's credential exchanges, the Terraform agent would not be able to authenticate with Auth0 (no scopes!).  This would also happen any time any token was forged, making debugging very difficult.  Adding the scopes to the token during the hook execution fixes this problem.